+++ pkg/resources/cluster/resources.go
@@ -2,5 +2,5 @@ package cluster
 
 import "embed"
 
-//go:embed satellite-common controller csi-controller csi-node satellite patches ha-controller
+//go:embed satellite-common controller csi-controller csi-node satellite patches ha-controller affinity-controller
 var Resources embed.FS